requirements.txt
mediapipe>=0.10.14
opencv-python>=4.9.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
joblib>=1.3.0
pyserial>=3.5
pytest>=8.0.0
